requirements.txt
discord.py
twilio
pytoml